// types/MyInputType.ts
export type MyInputType = {
    name: string;
    id: string;
    type: string;
    placeholder?: string;
    label?: string;
    pattern?: string;
    className?: string;
    max?: number;
    min?: number;
    maxLength?: number;
    minLength?: number;
    step?: number;
    dataTitle?: string;
    errorMessage?: string;
    required?: boolean;
};